also instrument expressions that build the futures
