+++ spandoc-attribute/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
-name = "spandoc"
-version = "0.1.3"
-authors = [ "Jane Lusby <jlusby@yaah.dev>" ]
+name = "spandoc-attribute"
+version = "0.1.0"
+authors = ["Jane Lusby <jlusby@yaah.dev>"]
+edition = "2018"
 repository = "https://github.com/yaahc/spandoc"
 homepage = "https://github.com/yaahc/spandoc"
-documentation = "https://docs.rs/spandoc"
+documentation = "https://docs.rs/spandoc-attribute"
 description = """
 Procedural macro attribute for converting doc comments into tracing spans
 """
@@ -16,14 +17,13 @@ categories = [
 keywords = ["logging", "tracing", "macro", "instrument", "log"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
-edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+proc-macro = true
+
 
 [dependencies]
-tracing = "0.1.15"
-tracing-futures = "0.2.4"
-spandoc-attribute = { path = "./spandoc-attribute" }
-
-[dev-dependencies]
-tracing-error = "0.1.2"
-tracing-subscriber = { features = ["registry", "fmt"], version = "0.2.6" }
-trybuild = "1.0.30"
+syn = { version = "1.0.33", features = ["full", "extra-traits", "fold"] }
+quote = "1.0.7"
+proc-macro2 = "1.0.18"
