+++ tests/compile/with_multi_await.rs
@@ -32,4 +32,5 @@ fn main() {
         let _four = fut.await.await;
     };
 
+
 }
